// src/interfaces/Metrics.ts
interface MetricsResponse {
    collections: number;
    custodianNetworks: number;
    custodians: number;
    datasets: number;
    datasetGmi: number;
    datasetCohortRequest: number;
    durs: number;
    publications: number;
    tools: number;
    users: number;
}

export type { MetricsResponse };